Lots of Fixes (Read Description)
- Fixed bug with appointments being null when getting appointment object from id
- Added setting maximum appointments to the settings page. You are not able to open the create an appointment prompt if you are at or above the max for the week. Still needs to be done on the appointment list page.
- Made it so if it is Saturday, the calendar's show next week, not the week that just past.

+++ app/controllers/Databases/AppointmentsDB.java
@@ -325,7 +325,7 @@ public class AppointmentsDB {
 
     public static AppointmentsModel removeAppointment(String appointmentId){
         AppointmentsModel appointment = getAppointment(appointmentId);
-        if(!appointment.getStartDate().before(new Date())){
+        if(!(appointment == null) && !appointment.getStartDate().before(new Date())){
             ApiFuture<WriteResult> writeResult = FirestoreDB.get().collection("appointments").document(appointmentId).delete();
         }
         /* Asynchronously remove appointment from DB */
